Change parent condition of stream-error on ABCL
ABCL has a strange quirk where a defined condition can declare
stream-error as a subtype, and a constructed condition of that type is typep to
stream-error, but stream-error-stream raises an error.

This change changes the parent of parser-error on ABCL from stream-error
to error, and adds the more portable function parser-error-stream to all
implementations.